[SPIKING] Move filtering classes to Utility
